refactor: move code into Focus.ts
to enable further refactoring
